+++ src/main/java/org/prgrms/nabimarketbe/domain/suggestion/dto/request/SuggestionRequestDTO.java
@@ -1,4 +1,4 @@
-package org.prgrms.nabimarketbe.domain.suggestion.dto;
+package org.prgrms.nabimarketbe.domain.suggestion.dto.request;
 
 public record SuggestionRequestDTO(Long fromCardId, Long toCardId) {
 }
