+++ src/main/java/org/prgrms/nabimarketbe/domain/suggestion/dto/response/SuggestionListReadPagingResponseDTO.java
@@ -1,0 +1,9 @@
+package org.prgrms.nabimarketbe.domain.suggestion.dto.response;
+
+import java.util.List;
+
+public record SuggestionListReadPagingResponseDTO(
+        List<SuggestionListReadResponseDTO> suggestionList,
+        String nextCursorId
+) {
+}
